+++ src/modules/graphics/graphics.h
@@ -196,6 +196,7 @@ void lovrBufferDestroy(void* ref);
 const BufferInfo* lovrBufferGetInfo(Buffer* buffer);
 void* lovrBufferGetData(Buffer* buffer, uint32_t offset, uint32_t extent);
 void* lovrBufferSetData(Buffer* buffer, uint32_t offset, uint32_t extent);
+void lovrBufferFlush(Buffer* buffer);
 bool lovrBufferCopy(Buffer* src, Buffer* dst, uint32_t srcOffset, uint32_t dstOffset, uint32_t extent);
 bool lovrBufferClear(Buffer* buffer, uint32_t offset, uint32_t extent, uint32_t value);